[#4188116] Implemented Morty's following comment in  List Views of Salesorder and Index page
The case of the path needs to match the case of the directory structure.

The markup for o-stage and o-stage-* are specific to opportunities and should not be on a sales order view. Also, sales orders do not contain a property called Stage.

The namespace for sales orders should be Mobile.SalesLogix.SalesOrder.