[#4188051] List view multi-select
List view didn't have a feature to select multiple items and perform an action
on them.

With this change, we have a checkbox with each list item, that can be checked
to perform an action. Float bar has an additional action button to show
selected items' key.

We have to move this feature to a super class, where it could be used by all the
views.