INFORCRM-21223
Fixing the error log list to detail transition. Removing the browsers alert from the Contour location search error handler. Adding the geolocation error to the error log, and displaying it to the user via a toast. The user will see a proper message from the browser as to why the geolocation is failing, instead of just our generic error message.