Opportunity Pricing (#440)
* Call repricing service for Opportunity

Add a new opportunity detail quick action for refreshing the opportunity
price. This calls the "RePriceOpportunity" business method.

Refs: INFORCRM-23307

* Disable the refresh pricing quick action based on integration settings

The opportunity detail quick action for refresh pricing will check
the integration settings for "Back Office Extension" and check if the
"Local CRM Pricing Opportunity" value is True or False. If the value is
not "True", the quick action will be disabled.

Refs: INFORCRM-23307