Keep the left nav/drawer search widget in sync with the list (#443)
If you initiate a speedsearch from the left drawer, the speedsearch list
will open and the search input at the top will match what the user input
on the left drawer. However, if you change the search term at the top
of the list view and do another search, the left drawer/nav becomes out
of sync and displays the previous search term instead. Moving the search
term to the redux store added in 4.0 to keep this state in sync.

Fixes: INFORCRM-2909