Fix removing the lead on a note when toggling "for lead" (#445)
* Fix removing the lead on a note when toggling "for lead"

Toggling the UI element would just hide the fields for the lead. When
creating the PUT request to do the update, the form calls getValues to
collect dirty values for the payload. Since only the UI was toggled, the
underlying field is not dirty. In this case, we can null them out by
overriding getValues in the activity/history edit form.

Fixes: INFORCRM-15723