Make prefix and suffix fields optional on contact name edit
The prefix and suffix fields were set to required. If the user clicked
on one of these lookups and selected a value, they would have to cancel
editing the entire form to undo the change. This change will make the
"None" option available on the lookup so it can be cleared.

Fixes: INFORCRM-8804