Load custom settings from server
Currently mobile stores most of its settings in configuration/*.js. This
requires an administrator to edit this file and re-deploy the mobile
portal to update them. In version 3.6 we added support for fetching
integration settings and loading them dynamically after authentication.
This will extend that to fetch custom settings for "Mobile". We can then
override the settings we loaded from the *.js files. If no custom
settings are loaded, it will continue using the config files.

Refs: INFORCRM-23508