Changes to contact list lookup for account name.
Search on the "AccountName" property rather than a joined column to
Account/AccountNameUpper. In addition to this, do not add the
Account/AccountName column to the query select in SData. It was not
being used in the list. Fixed contact search fields in help stating
that first and last name were being searched with "contains" when
actually they were searched with a "starts with".

Refs: INFORCRM-23884