Topic serviceworker (#455)
* Removing cache manifest and enableUpdateNotification code

Adding a stub for the service worker. Might move this to the SDK
later on.

Refs: INFORCRM-24403

* Add serviceworker.js to deployment. Add new configuration flag.

Added a new enableServiceWorker configuration flag. Service workers
will be enabled if offline or this new flag are true.

Refs: INFORCRM-24403

* Use new loadCache appmodule method for caching resources.

Added two image urls as a POC.
Refs: INFORCRM-24403

* Add application method to register multiple urls.

Refs: INFORCRM-24403

* Add some helper methods for checking service worker avail.

Cache localization files in bootstrap.
Refs: INFORCRM-24403

* Add cacheFiles array param to bootstrap.

Refs: INFORCRM-24403

* Updating list of files to cache for production.

Update index-body code behind to enumerate our static assets. This
will get passed to the service worker in bootstrap to cache. Adjust
the service worker to not send a cache response for ping.gif.

Refs: INFORCRM-24403

* Return a 408 timeout response when pinging offline.

Refs: INFORCRM-24403

* Remove test code.

* Skip bulk adding files if they are already in the cache.

Setup the bootstrap to not run the application until the promise
for cache registration completes.

Refs: INFORCRM-24403

* Always return a promise from registerCacheUrls.

* Added a toast when application has been cached by service worker.

Refs: INFORCRM-24403

* Clearing storage from settings will clear service worker caches

#Refs: INFORCRM-24403