Add environment based configuration system for e2e tests. (#463)
* Add environment based configuration system for e2e tests.

This will allow command line based runners to override these
settings without writing files.

* Fix activity category test selector issue when not in slowMo.

The activity edit form has category enabled by default. When the
base class calls setValues we conditionally check to enable/disable
it. This means for a split millisecond the field is not yet
disabled. This fixes the selector to wait for the to be in the
desired state by checking for [disabled] and :not([disabled]).

* Add e2e test README.md