Filter dropdowns in settings->language settings
The current language and regional dropdowns in language settings
are just loading based on the global languages available. This
currently is OK since we have available language strings and
regions in sync. However, if we were to go ahead and add regions
without strings, the mismatch will allow the user to select a
language that will not have strings translated. This changeset will
filter down the available list based on what the L20N context has
loaded.