Activity List Complete Quick Action Changes (#485)
We currently do not allow a quick action from list to complete an
activity. This change allows completing occurrences or the entire
series. The user will be given a prompt if the activity is
recurring. If the activity is not recurring, there will be no prompt
and the activity will just complete like previous versions

Refs:
INFORCRM-26164